requirements.txt
streamlit
numpy
scipy
soundfile
audioread
streamlit-knobs